bitaxe_sentry/requirements.txt
requests
apscheduler
sqlmodel
fastapi
uvicorn
python-dotenv
jinja2 